LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-single-value-in-shadow-tree-expected.txt


FAIL Meta color-scheme in shadow tree does not apply. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "auto"
